Break out some small third party libs to keep us on the straight and narrow
